LRAC-12616 Create '__tests__' folder to Components and move files to it
