+++ modules/apps/analytics/analytics-settings-web/src/main/resources/META-INF/resources/js/components/__tests__/Loading.spec.tsx
@@ -16,7 +16,7 @@ import '@testing-library/jest-dom/extend-expect';
 import {render, screen} from '@testing-library/react';
 import React from 'react';
 
-import Loading from './Loading';
+import Loading from '../Loading';
 
 describe('Loading', () => {
 	it('renders Loading component without crashing', async () => {
