+++ modules/apps/analytics/analytics-settings-web/src/main/resources/META-INF/resources/js/components/__tests__/StateRenderer.spec.tsx
@@ -19,7 +19,7 @@ import React from 'react';
 import StateRenderer, {
 	EmptyStateComponent,
 	ErrorStateComponent,
-} from './StateRenderer';
+} from '../StateRenderer';
 
 describe('State Renderer', () => {
 	it('renders EmptyStateComponent', () => {
